more refactor based on Brads suggestions
